Finish integration test 1
