+++ sdk/state/ingest.go
@@ -1,0 +1,84 @@
+package state
+
+import (
+	"fmt"
+
+	"github.com/stellar/go/txnbuild"
+)
+
+// IngestTx accepts any transaction that has been seen as successful or
+// unsuccessful on the network. The function updates the internal state of the
+// channel if the transaction relates to the channel.
+//
+// TODO: Signal when the state of the channel has changed to closed or closing.
+// TODO: Accept the xdr.TransactionResult and xdr.TransactionMeta so code can
+// determine if successful or not, and understand changes in the ledger as a
+// result.
+func (c *Channel) IngestTx(tx *txnbuild.Transaction) error {
+	// If the tx hash matches an unauthorized decl, copy the close signature.
+	err := c.ingestTxToUpdateUnauthorizedCloseAgreement(tx)
+	if err != nil {
+		return err
+	}
+
+	// TODO: If the tx hash matches an authorized or unauthorized declaration,
+	// mark the channel as closing.
+	// TODO: If the tx hash matches an authorized or unauthorized close, mark the
+	// channel as closed.
+	// TODO: If the tx is for an older declaration, mark the channel as closing with
+	// requiring bump.
+	// TODO: Use the transaction result to affect on success/failure.
+
+	return nil
+}
+
+func (c *Channel) ingestTxToUpdateUnauthorizedCloseAgreement(tx *txnbuild.Transaction) error {
+	if tx.SourceAccount().AccountID != c.initiatorEscrowAccount().Address.Address() {
+		return nil
+	}
+	ca := c.latestUnauthorizedCloseAgreement
+	if ca.isEmpty() {
+		return nil
+	}
+	txHash, err := tx.Hash(c.networkPassphrase)
+	if err != nil {
+		return fmt.Errorf("hashing tx: %w", err)
+	}
+	declTx, closeTx, err := c.closeTxs(c.openAgreement.Details, ca.Details)
+	if err != nil {
+		return fmt.Errorf("building txs for latest unauthorized close agreement: %w", err)
+	}
+	declTxHash, err := declTx.Hash(c.networkPassphrase)
+	if err != nil {
+		return fmt.Errorf("hashing latest unauthorized declaration tx: %w", err)
+	}
+	if txHash != declTxHash {
+		return nil
+	}
+	if err != nil {
+		return fmt.Errorf("building txs for latest unauthorized close agreement: %w", err)
+	}
+	closeTxHash, err := closeTx.Hash(c.networkPassphrase)
+	if err != nil {
+		return fmt.Errorf("hashing latest unauthorized close tx: %w", err)
+	}
+	for _, sig := range tx.Signatures() {
+		err = c.remoteSigner.Verify(declTxHash[:], sig.Signature)
+		if err == nil {
+			ca.ConfirmerSignatures.Declaration = sig.Signature
+			break
+		}
+	}
+	for _, sig := range tx.Signatures() {
+		err = c.remoteSigner.Verify(closeTxHash[:], sig.Signature)
+		if err == nil {
+			ca.ConfirmerSignatures.Close = sig.Signature
+			break
+		}
+	}
+	_, err = c.ConfirmPayment(ca)
+	if err != nil {
+		return fmt.Errorf("confirming the last unauthorized close: %w", err)
+	}
+	return nil
+}
